Fixed all compiler errors in the monte_carlo_collision_photon package.

+++ packages/monte_carlo/collision/photon/src/MonteCarlo_CoupledStandardCompleteDopplerBroadenedPhotonEnergyDistribution_def.hpp
@@ -61,7 +61,7 @@ double CoupledStandardCompleteDopplerBroadenedPhotonEnergyDistribution<ComptonPr
  */
 template<typename ComptonProfilePolicy>
 void CoupledStandardCompleteDopplerBroadenedPhotonEnergyDistribution<ComptonProfilePolicy>::sampleInteractionSubshell(
-                                               unsigned& old_subshell_index,
+                                               size_t& old_subshell_index,
                                                double& subshell_binding_energy,
                                                Data::SubshellType& subshell ) const
 {
